Security Hotspots First Review

+++ app/src/main/java/com/simplecity/amp_library/ui/screens/tagger/TaggerTask.java
@@ -169,12 +169,16 @@ public class TaggerTask extends AsyncTask<Object, Integer, Boolean> {
                 if (tempFiles != null && tempFiles.size() != 0) {
                     for (int j = tempFiles.size() - 1; j >= 0; j--) {
                         File file = tempFiles.get(j);
-                        try {
-                            java.nio.file.Files.delete(file.toPath());
-                        } catch (IOException e) {
-                            e.printStackTrace();
+                        if (BuildConfig.DEBUG) {
+                            try {
+                                java.nio.file.Files.delete(file.toPath());
+                            } catch (IOException e) {
+                                if (BuildConfig.DEBUG) {
+                                    e.printStackTrace();
+                                }
+                            }
+                            tempFiles.remove(j);
                         }
-                        tempFiles.remove(j);
                     }
                 }
             }
